move storage enum constants from top-level `constants.ts` to `storage/constants.ts`
